FE-1500: extract the simulation timeline
